+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add human-readable change dump.
+- Add `-t` option to simulate until a fixed number of steps have elapsed.
+
 ### Changed
 - Update llhd to v0.9.0.